# frontend/moderator-dashboard-ui/api-contracts.yaml
openapi: 3.1.0
info:
  title: B2B Platform API
  version: 1.0.0
paths:
...
  /moderator/parsing-runs:
    get:
      tags:
      - ModeratorTasks
      summary: List Parsing Runs
      operationId: list_parsing_runs_moderator_parsing_runs_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /moderator/parsing-runs/{runId}:
    get:
      tags:
      - ModeratorTasks
      summary: Get Parsing Run Detail
      operationId: get_parsing_run_detail_moderator_parsing_runs__runId__get
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParsingResultsResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /moderator/parsing-runs/{runId}/logs:
    get:
      tags:
      - ModeratorTasks
      summary: Get Parsing Run Logs
      operationId: get_parsing_run_logs_moderator_parsing_runs__runId__logs_get
      parameters:
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParsingRunLogResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /moderator/requests/{requestId}/start-parsing:
    post:
      tags:
      - ModeratorTasks
      summary: Start Parsing
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartParsingRequestDTO'
      description: 'Starts parsing for a request.

        MVP: request keys are mocked (later: fetch real keys from DB).'
      operationId: start_parsing_moderator_requests__requestId__start_parsing_post
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: integer
          title: Requestid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartParsingResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /moderator/suppliers:
    get:
      tags:
      - ModeratorSuppliers
      summary: List Moderator Suppliers
      operationId: list_moderator_suppliers_moderator_suppliers_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - ModeratorSuppliers
      summary: Create Moderator Supplier
      operationId: create_moderator_supplier_moderator_suppliers_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /moderator/suppliers/{supplierId}:
    get:
      tags:
      - ModeratorSuppliers
      summary: Get Moderator Supplier
      operationId: get_moderator_supplier_moderator_suppliers__supplierId__get
      parameters:
      - name: supplierId
        in: path
        required: true
        schema:
          type: integer
          title: Supplierid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModeratorSupplierDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - ModeratorSuppliers
      summary: Update Moderator Supplier
      operationId: update_moderator_supplier_moderator_suppliers__supplierId__put
      parameters:
      - name: supplierId
        in: path
        required: true
        schema:
          type: integer
          title: Supplierid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateModeratorSupplierRequestDTO'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModeratorSupplierDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /moderator/keywords:
    get:
      tags:
      - ModeratorKeywords
      summary: List Keywords Base
      operationId: list_keywords_moderator_keywords_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
          description: Substring search in keyword text
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            enum:
            - pending
            - parsed
            - failed
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: sort
        in: query
        required: false
        schema:
          type: string
          enum:
          - keyword_asc
          - keyword_desc
          - created_at_desc
          - created_at_asc
          default: keyword_asc
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeywordsListResponseDTO'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ParsingDomainGroupDTO:
      properties:
        domain:
          type: string
          title: Domain
        urls:
          items:
            type: string
          type: array
          title: Urls
        source:
          anyOf:
          - $ref: '#/components/schemas/ParsingSource'
          - type: 'null'
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
      type: object
      required:
      - domain
      - urls
      title: ParsingDomainGroupDTO
      description: 'SSoT: api-contracts.yaml#/components/schemas/ParsingDomainGroupDTO'
...
    ParsingResultsByKeyDTO:
      properties:
        keyId:
          type: integer
          title: Keyid
        groups:
          items:
            $ref: '#/components/schemas/ParsingDomainGroupDTO'
          type: array
          title: Groups
      type: object
      required:
      - keyId
      - groups
      title: ParsingResultsByKeyDTO
      description: 'SSoT: api-contracts.yaml#/components/schemas/ParsingResultsByKeyDTO'
    ParsingResultsResponseDTO:
      properties:
        requestId:
          type: integer
          title: Requestid
        runId:
          type: string
          title: Runid
        results:
          items:
            $ref: '#/components/schemas/ParsingResultsByKeyDTO'
          type: array
          title: Results
      type: object
      required:
      - requestId
      - runId
      - results
      title: ParsingResultsResponseDTO
      description: 'SSoT: api-contracts.yaml#/components/schemas/ParsingResultsResponseDTO'
    ParsingRunStatus:
      type: string
      enum:
      - queued
      - running
      - succeeded
      - failed
      title: ParsingRunStatus
      description: 'SSoT: api-contracts.yaml#/components/schemas/ParsingRunStatus'
    StartParsingRequestDTO:
      type: object
      additionalProperties: false
      properties:
        depth:
          anyOf:
          - type: integer
            minimum: 1
            maximum: 50
          - type: 'null'
          title: Depth
          description: How many search result pages to fetch per key. Default 10.
        source:
          anyOf:
          - $ref: '#/components/schemas/ParsingRunSource'
          - type: 'null'
          title: Source
          description: Which search engine(s) to use for this parsing run.
      title: StartParsingRequestDTO
      description: 'SSoT: api-contracts.yaml#/components/schemas/StartParsingRequestDTO'
    ParsingRunSource:
      type: string
      enum:
      - google
      - yandex
      - both
      title: ParsingRunSource
      description: 'Search engine selector for parsing run (request).'
    ParsingSource:
      type: string
      enum:
      - google
      - yandex
      title: ParsingSource
      description: 'SSoT: api-contracts.yaml#/components/schemas/ParsingSource'
...
    StartParsingResponseDTO:
      properties:
        requestId:
          type: integer
          title: Requestid
        runId:
          type: string
          title: Runid
        status:
          $ref: '#/components/schemas/ParsingRunStatus'
      type: object
      required:
      - requestId
      - runId
      - status
      title: StartParsingResponseDTO
      description: 'SSoT: api-contracts.yaml#/components/schemas/StartParsingResponseDTO'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError